+++ Demo/src/main/java/com/paypal/android/ui/card/DataCollectorHandler.kt
@@ -1,8 +1,8 @@
 package com.paypal.android.ui.card
 
 import android.content.Context
-import com.paypal.android.paypaldatacollector.PayPalDataCollector
-import com.paypal.android.paypaldatacollector.PayPalDataCollectorEnvironment
+import com.paypal.android.fraudprotection.PayPalDataCollector
+import com.paypal.android.fraudprotection.PayPalDataCollectorEnvironment
 import dagger.hilt.android.qualifiers.ApplicationContext
 import javax.inject.Inject